Added check for file or dir existance

+++ bin/src/filesystem/file/file.service.ts
@@ -1,4 +1,4 @@
-import { handleServerError, extractPath, FileSystemCommand } from '../../utils';
+import { handleServerError, extractPath, FileSystemCommand, isExists } from '../../utils';
 import { File } from '../models/file.model';
 import { Request, Response } from 'express';
 
@@ -11,6 +11,13 @@ export const remove = async (req: Request, res: Response) => {
     }
 
     const filePath = extractPath(req.path, userDir, FileSystemCommand.Remove);
+
+    const fileExists = await isExists(filePath);
+
+    if (!fileExists || req.path === FileSystemCommand.Remove) {
+      return res.status(403).json({ error: 'The path does not exist.' });
+    }
+
     await File.remove(filePath);
     return res.status(200).json({ message: 'File removed.' });
   } catch (error) {
@@ -25,6 +32,13 @@ export const download = async (req: Request, res: Response) => {
   }
 
   const filePath = extractPath(req.path, userDir, FileSystemCommand.Download);
+
+  const dirExists = await isExists(filePath);
+
+  if (!dirExists || req.path === FileSystemCommand.Download) {
+    return res.status(403).json({ error: 'The path does not exist.' });
+  }
+
   return res.download(filePath, async (error) => {
     await handleServerError(error, 500, res);
   });
